requirements.txt
asgiref==3.2.10
Django==3.0.8
django-crispy==1.8.1
django-crispy-forms==1.9.1
django-model-utils==4.0.0
django-notifications-hq==1.6.0
djangorestframework==3.11.0
jsonfield==3.1.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
swapper==1.1.2.post1
